+++ Function_Pointer/src/main.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <iostream>
-#include "func_ptr.hpp"
+#include "func_ptr/func_ptr.hpp"
 
 
 int main() {
